- Add gopls directory to gopls PATH - Extract path resolving to a separate class
