+++ ca.mt.gore/src/ca/mt/gore/GoServer.java
@@ -1,8 +1,6 @@
 package ca.mt.gore;
 
-import com.google.common.base.Splitter;
 import com.google.common.base.Strings;
-import com.google.common.io.ByteStreams;
 import com.google.gson.Gson;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
@@ -14,12 +12,8 @@ import java.io.InputStream;
 import java.io.OutputStream;
 import java.io.Reader;
 import java.net.URI;
-import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
-import java.util.function.Supplier;
 import org.eclipse.core.runtime.Platform;
 import org.eclipse.lsp4e.LanguageServerPlugin;
 import org.eclipse.lsp4e.server.ProcessOverSocketStreamConnectionProvider;
@@ -27,9 +21,38 @@ import org.eclipse.lsp4e.server.ProcessStreamConnectionProvider;
 import org.eclipse.lsp4e.server.StreamConnectionProvider;
 
 public class GoServer implements StreamConnectionProvider {
+	private final class ProcessStreamConnectionProviderExtension extends ProcessStreamConnectionProvider {
+		private ProcessStreamConnectionProviderExtension(List<String> commands, String workingDir) {
+			super(commands, workingDir);
+		}
+
+		@Override
+		protected ProcessBuilder createProcessBuilder() {
+			ProcessBuilder pb = super.createProcessBuilder();
+			addGoplsLocationToEnvironment(pb.environment());
+			return pb;
+		}
+	}
+
+	private final class ProcessOverSocketStreamConnectionProviderExtension
+			extends ProcessOverSocketStreamConnectionProvider {
+		private ProcessOverSocketStreamConnectionProviderExtension(List<String> commands, String workingDir, int port) {
+			super(commands, workingDir, port);
+		}
+
+		@Override
+		protected ProcessBuilder createProcessBuilder() {
+			ProcessBuilder pb = super.createProcessBuilder();
+			addGoplsLocationToEnvironment(pb.environment());
+			return pb;
+		}
+	}
+
 	private static final int CONNECTION_PORT = 4389;
 
 	private StreamConnectionProvider provider;
+
+	private String goplsDir;
 
 	public GoServer() {}
 
@@ -93,84 +116,33 @@ public class GoServer implements StreamConnectionProvider {
 
 	private void createProvider() throws IOException {
 		String os = Platform.getOS();
-		List<String> commands = new ArrayList<>();
 
-		String gopls = resolveGoBinary("gopls");
-		if (gopls != null) {
-			LanguageServerPlugin.logInfo("GORE: Found gopls at " + gopls);
-			commands.add(gopls);
-			commands.add("serve");
-			// commands.add("-rpc.trace");
-			if (os.equals(Platform.OS_WIN32)) {
-				commands.add("--listen=127.0.0.1:" + CONNECTION_PORT);
-			}
-		}
-
-		if (commands.isEmpty()) {
+		String gopls = GoResolver.resolveGoBinary("gopls");
+		if (gopls == null) {
 			IOException ex = new IOException("Cannot find gopls");
 			LanguageServerPlugin.logError("GORE: Could not find gopls or bingo", ex);
 			throw ex;
 		}
 
+		LanguageServerPlugin.logInfo("GORE: Found gopls at " + gopls);
+		goplsDir = gopls.substring(0, gopls.lastIndexOf(File.separator));
+
+		List<String> commands = List.of(gopls, "serve"); // "-rpc.trace"
 		File workingDir = new File(".");
 		if (os.equals(Platform.OS_WIN32)) {
-			provider = new ProcessOverSocketStreamConnectionProvider(commands, workingDir.toString(),
-					CONNECTION_PORT) {};
+			commands.add("--listen=127.0.0.1:" + CONNECTION_PORT);
+			provider = new ProcessOverSocketStreamConnectionProviderExtension(commands, workingDir.toString(),
+					CONNECTION_PORT);
 		} else {
-			provider = new ProcessStreamConnectionProvider(commands, workingDir.toString()) {};
+			provider = new ProcessStreamConnectionProviderExtension(commands, workingDir.toString());
 		}
 	}
 
-	private static String resolveGoBinary(String program) {
-		if (Platform.OS_WIN32.equals(Platform.getOS()) && !program.toLowerCase().endsWith(".exe")) {
-			program += ".exe";
-		}
-		List<Supplier<String>> pathProviders = Arrays.asList(() -> System.getenv("GOBIN"),
-				() -> System.getenv("GOPATH"), () -> run("go", "env", "GOBIN"), () -> run("go", "env", "GOPATH"),
-				() -> new File(System.getenv("HOME"), "go").toString());
-		for (Supplier<String> provider : pathProviders) {
-			String path = provider.get();
-			if (!Strings.isNullOrEmpty(path)) {
-				String fullPath = resolve(path, program);
-				if (fullPath != null) {
-					return fullPath;
-				}
-				fullPath = resolve(path, "bin" + File.separator + program);
-				if (fullPath != null) {
-					return fullPath;
-				}
-			}
-		}
-
-		return resolve(System.getenv("PATH"), program);
-	}
-
-	private static String run(String... cmd) {
-		try {
-			Process process = new ProcessBuilder(cmd).start();
-			if (process.waitFor() != 0) {
-				return null;
-			}
-			return new String(ByteStreams.toByteArray(process.getInputStream()), StandardCharsets.UTF_8).trim();
-		} catch (IOException ex) {
-			return null;
-		} catch (InterruptedException ex) {
-			return null;
-		}
-	}
-
-	private static String resolve(String env, String relativeFile) {
-		if (Strings.isNullOrEmpty(env)) {
-			return null;
-		}
-		env = env.trim();
-		for (String path : Splitter.on(File.pathSeparator).split(env)) {
-			File file = new File(path, relativeFile);
-			if (file.exists() && file.canExecute()) {
-				return file.getPath();
-			}
-		}
-		return null;
+	/**
+	 * Update the gopls environment by adding the gopls location to the PATH.
+	 */
+	public void addGoplsLocationToEnvironment(Map<String, String> environment) {
+		environment.merge("PATH", goplsDir, (pathValue, update) -> pathValue + File.pathSeparator + update);
 	}
 
 	@Override
